fix: MOO and moo not jumping to the correct command
Assuming MOO is the 'start loop' label and 'moo' is the 'end loop' label, the bug consists in both instructions not ignoring MOO and 'moo' commands that make nested loops. This means a MOO command could jump at the end of a nested loop instead of jumping at the end of the outer loop. The reasoning is the same for the 'moo' command.
This wasn't easy to find, now the 99bottles program gives the expected output.